Image save and get completed 99%

+++ src/main/java/com/ph/service/ImageService.java
@@ -5,15 +5,15 @@ package com.ph.service;
 import com.ph.exception.customs.ImageException;
 import com.ph.exception.customs.ResourceNotFoundException;
 import com.ph.payload.mapper.ImageMapper;
- import com.ph.payload.request.ImageRequest;
+
  import com.ph.payload.response.ImageResponse;
 import com.ph.repository.AdvertRepository;
 import com.ph.repository.ImageRepository;
 import com.ph.utils.ImageUtil;
+ import com.ph.utils.MessageUtil;
  import jakarta.transaction.Transactional;
  import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
+
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
@@ -29,6 +29,7 @@ public class ImageService {
     private final ImageRepository imageRepository;
     private final ImageMapper imageMapper;
     private final AdvertRepository advertRepository;
+    private final MessageUtil messageUtil;
 
 
     public Image saveImage(MultipartFile request ,Long advertId){
@@ -40,7 +41,7 @@ public class ImageService {
                     .type(request.getContentType())
                     .data(ImageUtil.compressImage(request.getBytes()))
                     .advert(advertRepository.findById(advertId)
-                            .orElseThrow(() -> new ResourceNotFoundException(String.format("Advert not found with id: %s", advertId))))
+                            .orElseThrow(() -> new ResourceNotFoundException(String.format(messageUtil.getMessage("error.advert.not.found"), advertId))))
                     .build();
         } catch (IOException e) {
             throw new ImageException(e.getMessage());
@@ -51,9 +52,17 @@ public class ImageService {
 
 
 
-    public ImageResponse createImage(List<MultipartFile> request, ImageRequest advert  )   {
+    public ImageResponse createImage(List<MultipartFile> request, Long advertId  )   {
 
-  List<Image> savedImage =  request.stream().map(t->saveImage(t,advert.getAdvertId())).toList();
+
+
+    if (request.isEmpty()) {
+    throw new ImageException(String.format(messageUtil.getMessage("error.image.not.found")));
+    }
+
+
+
+  List<Image> savedImage =  request.stream().map(t->saveImage(t, advertId)).toList();
 
 
         return imageMapper.toImageResponse(savedImage.get(0));
@@ -63,7 +72,7 @@ public class ImageService {
 
     public ResponseEntity<?> getImageById(Long id) {
         Image found = imageRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException(String.format("Image not found with id: %s", id)));
+                .orElseThrow(() -> new ResourceNotFoundException(String.format(messageUtil.getMessage("error.image.not.found.id"), id)));
 
 
 
@@ -78,6 +87,12 @@ public class ImageService {
     }
 
     public void deleteImage(List<Long> id) {
+      Advert advert=  advertRepository.findByImages_Id(id.get(0));
+
+      if (advert.getImages().size()==id.size()){
+          throw new ImageException(String.format(messageUtil.getMessage("error.image.not.delete")));
+      }
+
         imageRepository.deleteAllById(id);
     }
 
@@ -85,17 +100,17 @@ public class ImageService {
     @Transactional
     public ImageResponse updateImage(Long imgId, Long advertId) {
       Advert advert =  advertRepository.findById(advertId)
-                .orElseThrow(() -> new ResourceNotFoundException(String.format("Advert not found with id: %s", advertId)));
+                .orElseThrow(() -> new ResourceNotFoundException(String.format(messageUtil.getMessage("error.advert.not.found"), advertId)));
 
       List<Image> advertImages = advert.getImages();
 
         boolean match = advertImages.stream().anyMatch(image -> image.getId().equals(imgId));
         if (!match) {
-            throw new ResourceNotFoundException(String.format("Image not found with id: %s", imgId));
+            throw new ResourceNotFoundException(String.format(messageUtil.getMessage("error.image.not.found.id"), imgId));
         }
 
         Image found = imageRepository.findById(imgId)
-                .orElseThrow(() -> new ResourceNotFoundException(String.format("Image not found with id: %s", imgId)));
+                .orElseThrow(() -> new ResourceNotFoundException(String.format(messageUtil.getMessage("error.image.not.found.id"), imgId)));
 
 
          Image featuredImage =   advertImages.stream().filter(Image::isFeatured).findFirst().orElse(null);
@@ -121,7 +136,7 @@ public class ImageService {
     @Transactional
     public List<ImageResponse> getAllImagesByAdvertId(Long id) {
         Advert advert =  advertRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException(String.format("Advert not found with id: %s", id)));
+                .orElseThrow(() -> new ResourceNotFoundException(String.format(messageUtil.getMessage("error.advert.not.found"), id)));
         List<Image> advertImages = advert.getImages();
 
         advertImages.sort(Comparator.comparing(Image::isFeatured).reversed());
